APIs/API_UpdateBD/requirements.txt
fastapi
uvicorn
cassandra-driver
pika
pydantic